Remove two casts that clang-tidy flags as redundant
The previous commit dropped the `const UInt64 d = dim;` alias in
`HierarchicalKMeansData::merge` so the function could use the `UInt32`
member directly. That turned `static_cast<UInt32>(d)`, which used to
narrow from `UInt64`, into a cast of a value that is already `UInt32`,
and `readability-redundant-casting` rejects it:

  AggregateFunctionHierarchicalKMeans.cpp:871:55: error: redundant explicit
  casting to the same type 'UInt32' (aka 'unsigned int') as the
  sub-expression, remove this casting
  [readability-redundant-casting,-warnings-as-errors]

Drop both casts. No behaviour change.

Verified with `clang-tidy-21 -p build` on both files of the pull request:
no diagnostics. Re-inserting one cast reproduces the error above, so the
clean run is not clang-tidy skipping the file.

CI report: https://s3.amazonaws.com/clickhouse-test-reports/praktika.html?PR=112309&sha=51be328f3edf98375c21d43eb2a1562379bc9219&name_0=PR
Related: https://github.com/ClickHouse/ClickHouse/pull/112309

+++ src/AggregateFunctions/AggregateFunctionHierarchicalKMeans.cpp
@@ -868,7 +868,7 @@ struct HierarchicalKMeansData
         if (other.seen <= cap)
         {
             for (UInt64 slot = 0; slot < have_theirs; ++slot)
-                addVector(&other.samples[slot * dim], static_cast<UInt32>(dim), cap);
+                addVector(&other.samples[slot * dim], dim, cap);
             return;
         }
 
@@ -882,7 +882,7 @@ struct HierarchicalKMeansData
             seen = other.seen;
             rng = other.rng;
             for (UInt64 slot = 0; slot < have_ours; ++slot)
-                addVector(&ours[slot * dim], static_cast<UInt32>(dim), cap);
+                addVector(&ours[slot * dim], dim, cap);
             return;
         }
